# [stdlib] Optimize `String.write()`