--- a/mojo/stdlib/benchmarks/collections/bench_string.mojo
+++ b/mojo/stdlib/benchmarks/collections/bench_string.mojo
@@ -283,6 +283,40 @@ fn bench_write_utf8[
 
 
 # ===-----------------------------------------------------------------------===#
+# Benchmark string write
+# ===-----------------------------------------------------------------------===#
+@parameter
+fn bench_string_write[short: Bool](mut b: Bencher) raises:
+    var items = make_string[1000]("UN_charter_EN.txt")
+    # workaround for "allows writing to mem location ..."
+    # even though I tried using an immutable StringSlice
+    var items_2 = items.copy()
+    var items_3 = items.copy()
+    var items_4 = items.copy()
+    var items_5 = items.copy()
+
+    @always_inline
+    @parameter
+    fn call_fn() raises:
+        for _ in range(1_000_000):
+            var res: String
+
+            @parameter
+            if short:  # less than 24 bytes
+                res = String.write(0, " is ", "a", String(" number"))
+            else:  # 5001 bytes long
+                res = String.write(0, items, items_2, items_3, items_4, items_5)
+            keep(Bool(res))
+
+    b.iter[call_fn]()
+    keep(Bool(items))
+    keep(Bool(items_2))
+    keep(Bool(items_3))
+    keep(Bool(items_4))
+    keep(Bool(items_5))
+
+
+# ===-----------------------------------------------------------------------===#
 # Benchmark Main
 # ===-----------------------------------------------------------------------===#
 def main():
@@ -325,6 +359,12 @@ def main():
     """
 
     m.bench_function[bench_string_init](BenchId("bench_string_init"))
+    m.bench_function[bench_string_write[True]](
+        BenchId(String("bench_string_write_short"))
+    )
+    m.bench_function[bench_string_write[False]](
+        BenchId(String("bench_string_write_long"))
+    )
 
     @parameter
     for i in range(len(lengths)):