Optimize `String.write()`

Inspired by #4562

@msaelices it's pretty much the same implementation I showed in https://github.com/modular/modular/pull/4562#issuecomment-2878160467 plus an end character. We should maybe take the join implementation to a `String.write` overload at some point

## Benchmark results

CPU: Intel® Core™ i7-7700HQ

|| old value (ms) | new value (ms) | percentage | magnitude|
|-- | -- | -- | -- | --|
|bench_string_write_short | 41.65 | 40.4554731666667 | 0.0286284878335058 | 1.02947223330614|
|bench_string_write_long | 451.978725333333 | 281.21740075 | 0.377808323737798 | 1.60722175842575|
|average improvement  | - |- | 0.203218405785652 | 1.31834699586594|