@JoeLoser 2025-06-04
This is still failing an internal test on H100 that needs investigation:

```
FAIL: //open-source/max/max/kernels/test/gpu/numerics:test_e4m3fn_conversion.mojo.test (Exit 1) (see /home/runner/.cache/bazel/_bazel_runner/h100/execroot/_main/bazel-out/k8-opt/testlogs/open-source/max/max/kernels/test/gpu/numerics/test_e4m3fn_conversion.mojo.test/test.log)
INFO: From Testing //open-source/max/max/kernels/test/gpu/numerics:test_e4m3fn_conversion.mojo.test:
==================== Test output for //open-source/max/max/kernels/test/gpu/numerics:test_e4m3fn_conversion.mojo.test:
open-source/max/max/kernels/test/gpu/numerics/test_e4m3fn_conversion.mojo:573:10: error: CHECK: expected string not found in input
# CHECK: 0.0, 0.001953125, 0.00390625, 0.005859375, 0.0078125, 0.009765625, 0.01171875, 0.013671875,
         ^
<stdin>:198:34: note: scanning from here
== test_simd_e4m3_to_f16_ptx_path
                                 ^
<stdin>:199:1: note: possible intended match here
0.0, 0.001953125, 0.00390625, 0.005859375, 0.0078125, , 0.009765625, 0.01171875, 0.013671875, 
^
open-source/max/max/kernels/test/gpu/numerics/test_e4m3fn_conversion.mojo:620:10: error: CHECK: expected string not found in input
# CHECK: 0.0, 0.001953125, 0.00390625, 0.005859375, 0.0078125, 0.009765625, 0.01171875, 0.013671875,
         ^
<stdin>:231:41: note: scanning from here
an, 0, == test_simd_e4m3_to_f32_ptx_path
                                        ^
<stdin>:232:6: note: possible intended match here
0.0, 0, 0.001953125, 0.00390625, 0.005859375, 0.0078125, , 0.009765625, 0.01171875, 0.013671875, 
     ^

Input file: <stdin>
Check file: open-source/max/max/kernels/test/gpu/numerics/test_e4m3fn_conversion.mojo

-dump-input=help explains the following input dump.

Input was:
<<<<<<
             .
             .
             .
           193: 224.0, 224.0, 224.0, 224.0, 224.0, 224.0, 224.0, 224.0,  
           194: 224.0, 224.0, 224.0, 224.0, 224.0, 224.0, 224.0, 224.0,  
           195: 224.0, 240.0, 240.0, 240.0, 240.0, 240.0, 240.0, 240.0,  
           196: 240.0, 240.0, 240.0, 240.0, 240.0, 240.0, 240.0, 240.0,  
           197: 256.0, 256.0, 256.0, 256.0, 256.0, 256.0, 256.0, 256.0,  
           198: == test_simd_e4m3_to_f16_ptx_path 
check:573'0                                      X error: no match found
           199: 0.0, 0.001953125, 0.00390625, 0.005859375, 0.0078125, , 0.009765625, 0.01171875, 0.013671875,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
check:573'1     ?                                                                                               possible intended match
           200: .013671875, 0.015625, 5, 0.017578125, 0.01953125, 0.021484375, 0.0234375, , 0.0[253](https://github.com/modularml/modular/actions/runs/15117951544/job/42493282148?pr=61877#step:7:254)90625, 0.02734375, 0.029296875,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           201: .029296875, 0.03125, 75, 0.03515625, 0.0390625, 0.04296875, 0.046875, , 0.05078125, 0.0546875, 0.05859375,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           202: .05859375, 0.0625, 75, 0.0703125, 0.078125, 0.0859375, 0.09375, , 0.1015625, 0.109375, 0.1171875,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           203: .1171875, 0.125, 75, 0.140625, 0.15625, 0.171875, 0.1875, , 0.203125, 0.21875, 0.234375,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           204: .234375, 0.25, 75, 0.28125, 0.3125, 0.34375, 0.375, , 0.40625, 0.4375, 0.46875,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
             .
             .
             .
           226: 15.0, -16.0, -18.0, -20.0, -22.0, -24.0, -26.0, -28.0, -30.0,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           227: 30.0, -32.0, -36.0, -40.0, -44.0, -48.0, -52.0, -56.0, -60.0,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           228: 60.0, -64.0, -72.0, -80.0, -88.0, -96.0, -104.0, -112.0, -120.0,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           229: 120.0, -128.0, -144.0, -160.0, -176.0, -192.0, -208.0, -224.0, -240.0,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           230: 240.0, -[256](https://github.com/modularml/modular/actions/runs/15117951544/job/42493282148?pr=61877#step:7:257).0, -288.0, -320.0, -352.0, -384.0, -416.0, -448.0, nan, 0,  
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           231: an, 0, == test_simd_e4m3_to_f32_ptx_path 
check:573'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
check:620'0                                             X error: no match found
           232: 0.0, 0, 0.001953125, 0.00390625, 0.005859375, 0.0078125, , 0.009765625, 0.01171875, 0.013671875,  
check:620'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
check:620'1          ?                                                                                             possible intended match
           233: .013671875, 0.015625, 5, 0.017578125, 0.01953125, 0.021484375, 0.0234375, , 0.025390625, 0.0[273](https://github.com/modularml/modular/actions/runs/15117951544/job/42493282148?pr=61877#step:7:274)4375, 0.029296875,  
check:620'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           234: .029296875, 0.03125, 75, 0.03515625, 0.0390625, 0.04296875, 0.046875, , 0.05078125, 0.0546875, 0.05859375,  
check:620'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           235: .05859375, 0.0625, 75, 0.0703125, 0.078125, 0.0859375, 0.09375, , 0.1015625, 0.109375, 0.1171875,  
check:620'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           236: .1171875, 0.125, 75, 0.140625, 0.15625, 0.171875, 0.1875, , 0.203125, 0.21875, 0.234375,  
check:620'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
           237: .234375, 0.25, 75, 0.[281](https://github.com/modularml/modular/actions/runs/15117951544/job/42493282148?pr=61877#step:7:282)25, 0.3125, 0.34375, 0.375, , 0.40625, 0.4375, 0.46875,  
check:620'0     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
             .
             .
             .
>>>>>>
```

Separately, @jackos is reworking few of the writer/writable things and this may fall out for free/"just work?